+++ src/store/configureStore.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 import rootReducer from '../reducers/reducers';
-import sagas from '../sagas';
+import sagas from '../sagas/sagas';
 
 const configureStore = () => {
   const sagaMiddleWare = createSagaMiddleware();
